requirements.txt
IPython==8.26.0
openai==1.41.1
matplotlib==3.9.2
pandas==2.2.2
seaborn==0.13.2
oss2==2.18.6
python-dotenv==1.0.1
pymysql==1.1.1
google-api-python-client==2.141.0
google_auth_oauthlib==1.2.1
bs4==0.0.2
tiktoken==0.7.0
lxml==5.3.0
html2text==2024.2.26
nbconvert==7.16.4


fastapi==0.112.1
uvicorn==0.30.6